+++ front/app/containers/ProjectsShowPage/timeline/PhaseDocumentAnnotation/index.tsx
